CellArrays

Documentation for CellArrays.

CellArrays.CellArrays

CellArrays.CPUCellArray

CellArrays.CellArray

CellArrays.CuCellArray

CellArrays.ROCCellArray

CellArrays.blocklength

CellArrays.cellsize

CellArrays.CellArrays — Module

Module CellArrays

Provides support for an AbstractArray subtype CellArray, which represents arrays with cells that can contain logical arrays instead of single values; the data is stored in an optimal fashion for GPU HPC applications.

General overview and examples

https://github.com/omlins/CellArray.jl

Constructors

CellArray

CPUCellArray

CuCellArray

ROCCellArray

Functions (additional to standard AbstractArray functionality)

cellsize

blocklength

To see a description of a constructor or function type ?<constructorname> or ?<functionname>, respectively.

Performance note

If a CellArray's cells contain more than what fits into registers, the performance on Nvidia GPUs will deviate from the optimal, if access is not performed by accessing the cells' values individually to reduce register pressure.

CellArrays.CPUCellArray — Type

CPUCellArray{T<:Cell,N,B,T_elem} <: AbstractArray{T,N} where Cell <: Union{Number, SArray, FieldArray}

N-dimensional CellArray with cells of type T, blocklength B, and T_array being an Array of element type T_elem: alias for CellArray{T,N,B,Array{Telem,CellArrays.N}}.

CellArrays.CellArray — Type

CellArray{T<:Cell,N,B,T_array} <: AbstractArray{T,N} where Cell <: Union{Number, SArray, FieldArray}

N-dimensional array with elements of type T, where T are Cells of type Number, SArray or FieldArray. B defines the blocklength, which refers to the amount of values of a same Cell component that are stored contigously (B=1 means array of struct like storage; B=prod(dims) means array struct of array like storage; B=0 is an alias for B=prod(dims), enabling better peformance thanks to more specialized dispatch). T_array defines the array type used for storage.

CellArray{T,N,B}(T_arraykind, undef, dims)
CellArray{T,B}(T_arraykind, undef, dims)
CellArray{T}(T_arraykind, undef, dims)

Construct an uninitialized N-dimensional CellArray containing Cells of type T which are stored in an array of kind T_arraykind.

CPUCellArray{T,B}(undef, dims)
CPUCellArray{T}(undef, dims)

CuCellArray{T,B}(undef, dims)
CuCellArray{T}(undef, dims)

ROCCellArray{T,B}(undef, dims)
ROCCellArray{T}(undef, dims)

Construct an uninitialized N-dimensional CellArray containing Cells of type T which are stored in an array of kind Array, CuArray or ROCArray depending on the constructor chosen (CPUCellArray or CuCellArray or ROCCellArray) .

Performance note

Best performance on GPUs is in general obtained with B=0 as set by default. B=1 migth give better performance in certain cases. Other values of B do with the current implementation not lead to optimal performance on GPU.

CellArrays.CuCellArray — Type

CuCellArray{T<:Cell,N,B,T_elem} <: AbstractArray{T,N} where Cell <: Union{Number, SArray, FieldArray}

N-dimensional CellArray with cells of type T, blocklength B, and T_array being a CuArray of element type T_elem: alias for CellArray{T,N,B,CuArray{Telem,CellArrays.N}}.

CellArrays.ROCCellArray — Type

ROCCellArray{T<:Cell,N,B,T_elem} <: AbstractArray{T,N} where Cell <: Union{Number, SArray, FieldArray}

N-dimensional CellArray with cells of type T, blocklength B, and T_array being a ROCArray of element type T_elem: alias for CellArray{T,N,B,ROCArray{Telem,CellArrays.N}}.

CellArrays.blocklength — Method

blocklength(A)

Return the blocklength of CellArray A.

CellArrays.cellsize — Method

cellsize(A)
cellsize(A, dim)

Return a tuple containing the dimensions of A or return only a specific dimension, specified by dim.